+++ Basil/include/Renderer/Shader.h
@@ -13,7 +13,8 @@ namespace Basil
 			virtual ~Shader() {}
 			virtual void bind() const = 0;
 			virtual void unbind() const = 0;
-			static Shader* create(const std::string& filePath);
-			static Shader* create(const std::string& vertexSource, const std::string& fragmentSource);
+			virtual const std::string& getName() const = 0;
+			static Shared<Shader> create(const std::string& filePath);
+			static Shared<Shader> create(const std::string& name, const std::string& vertexSource, const std::string& fragmentSource);
 	};
 }