+++ Basil/include/pch.h
@@ -3,6 +3,7 @@
  */
 #pragma once
 
+#include <array>
 #include <functional>
 #include <iostream>
 #include <memory>